Create a new react app with create-react-app
Why is this change needed?
--------------------------
I wanted to create a new react app to ensure my understanding of the
basics of React as they exist today.

How does it address the issue?
------------------------------
In order to get up and running quickly, this generates a new React app
using create-react-app and the following command.
```
npx create-react-app .
```
The output of that command includes all the basic pieces needed to build
a React app.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://create-react-app.dev