Adds basic screen and details components
Why is this change needed?
--------------------------
We're starting by building out a static version of the UI in a few
phases. This will allow the entire UI to be in place before focusing on
functionality and lower level implementation details. Building out a
static version of the components also helps quickly identify component
architecture and where state should live.

How does it address the issue?
------------------------------
Starting at the top and working down, this commit add a couple of
components to the bottom half of the <Pokedex />:
  1. A <Screen /> component that will display an image and the <Details />
     of a given pokemon.
  2. A <Details /> component that includes some basic data about a
     Pokemon. This currently only includes name, type, number, and
     description, but we might include addition data points later.

Test files have been created, but currently only serve to ensure that
components successfully render. The typography at this point is rough
with the intention of adding polish later.