Populates pokemon data from the Pokemon API
Why is this change needed?
--------------------------
We need to do two basic things before we can make the search functionality
work. First, we request real data from an API to populate our components.
Second, we need to store a search string to use in the API query.

How does it address the issue?
------------------------------
This commit begins by accomplishing two things:
  1. It configures a GraphQL client to be utilized to make requests
     throughout the app.
  2, It creates a query.js file that contains a GraphQL query that will
     search pokemon by name using a partial match regex.

Using React hooks, we can fire off the query mentioned above and store
the result of the query as a state object on the <Pokemon /> component.
We can then pass that pokemon state object down to where the data is
needed.

by using the useLazyQuery function in conjunction with useEffect we can
send a request to the Pokemon API any time that data changes.
This specific useEffect function will only fire if the search text and
the data returned from the api changes or the API return a different set
of data.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://pokeapi.co/docs/graphql
https://reactjs.org/docs/hooks-effect.html