Dynamically loads pokemon image into the screen component
Why is this change needed?
--------------------------
We want to display the image of a pokemon when the API returns a result.
If no pokemon is found we want to display a fallback image.

How does it address the issue?
------------------------------
When a pokemon is passed into the <Screen /> component it will use the
id to load an image from a GitHub repo. While this isn't particularly
reliable, it will work for our demonstration purposes. If a pokemon is
not found, we will fallback to a placeholder image that looks like a
grayed out Bulbasaur.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://github.com/PokeAPI/sprites/tree/master/sprites/pokemon/other/dream-world