Request pokemon data based on search input value
Why is this change needed?
--------------------------
We want to actually be able to search for pokemon and display the data
that's returned from the API.

How does it address the issue?
------------------------------
The <Search /> component manages the input state internally. By sending
a setSearch function down to the <Search /> component we can track when
that value updates from the <Pokedex />. The setSearch function is
throttled to prevent extra API requests.