Cleans up overall UI design
Why is this change needed?
--------------------------
The Pokedex is now functional, but the UI could use some cleanup. Let's
make sure the typography looks nice and make sure that it works on
mobile devices.

How does it address the issue?
------------------------------
This swaps out the pixelated font for Roboto to make it feel more
Material Design inspired. The styles for the top of the pokedex have
also been tweaked to provide a better experience on various screen
sizes.