Removes limit on search query
Why is this change needed?
--------------------------
Let's load up a random Pokemon every time a the search box is empty.

How does it address the issue?
------------------------------
This removes the limit from the search query so we get back an array of
all the pokemon. If there is more than one returned from the search,
a random item will be selected. Along with this, the URL we get images
from has been updated to one that includes a larger number of images.
This was due to the lack of completeness of the svg images.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
https://github.com/PokeAPI/sprites/tree/master/sprites/pokemon/other/official-artwork